`Shamrock`: 绕过资源上传检测

+++ xposed/src/main/java/moe/fuqiuluo/qqinterface/servlet/transfile/NtV2RichMediaSvc.kt
@@ -1,14 +1,11 @@
 package moe.fuqiuluo.qqinterface.servlet.transfile
 
-import android.graphics.Bitmap
 import android.graphics.BitmapFactory
-import android.media.MediaMetadataRetriever
 import androidx.exifinterface.media.ExifInterface
 import com.tencent.mobileqq.qroute.QRoute
 import com.tencent.qqnt.aio.adapter.api.IAIOPttApi
 import com.tencent.qqnt.kernel.nativeinterface.CommonFileInfo
 import com.tencent.qqnt.kernel.nativeinterface.Contact
-import com.tencent.qqnt.kernel.nativeinterface.FileElement
 import com.tencent.qqnt.kernel.nativeinterface.MsgConstant
 import com.tencent.qqnt.kernel.nativeinterface.MsgElement
 import com.tencent.qqnt.kernel.nativeinterface.PicElement
@@ -16,28 +13,19 @@ import com.tencent.qqnt.kernel.nativeinterface.PttElement
 import com.tencent.qqnt.kernel.nativeinterface.QQNTWrapperUtil
 import com.tencent.qqnt.kernel.nativeinterface.RichMediaFilePathInfo
 import com.tencent.qqnt.kernel.nativeinterface.VideoElement
-import com.tencent.qqnt.msg.api.IMsgUtilApi
 import kotlinx.atomicfu.atomic
-import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.suspendCancellableCoroutine
-import kotlinx.coroutines.withContext
 import kotlinx.coroutines.withTimeoutOrNull
 import moe.fuqiuluo.qqinterface.servlet.BaseSvc
 import moe.fuqiuluo.qqinterface.servlet.TicketSvc
-import moe.fuqiuluo.qqinterface.servlet.transfile.data.Private
-import moe.fuqiuluo.qqinterface.servlet.transfile.data.Troop
 import moe.fuqiuluo.qqinterface.servlet.transfile.data.TryUpPicData
-import moe.fuqiuluo.qqinterface.servlet.transfile.data.VideoResource
-import moe.fuqiuluo.shamrock.helper.Level
 import moe.fuqiuluo.shamrock.helper.LogCenter
 import moe.fuqiuluo.shamrock.helper.MessageHelper
-import moe.fuqiuluo.shamrock.helper.TransfileHelper
 import moe.fuqiuluo.shamrock.remote.service.config.ShamrockConfig
 import moe.fuqiuluo.shamrock.tools.hex2ByteArray
 import moe.fuqiuluo.shamrock.tools.slice
 import moe.fuqiuluo.shamrock.utils.AudioUtils
 import moe.fuqiuluo.shamrock.utils.FileUtils
-import moe.fuqiuluo.shamrock.utils.MD5
 import moe.fuqiuluo.shamrock.utils.MediaType
 import moe.fuqiuluo.shamrock.xposed.helper.NTServiceFetcher
 import moe.fuqiuluo.shamrock.xposed.helper.msgService
@@ -63,7 +51,6 @@ import protobuf.oidb.cmd0x388.Cmd0x388ReqBody
 import protobuf.oidb.cmd0x388.Cmd0x388RspBody
 import protobuf.oidb.cmd0x388.TryUpImgReq
 import java.io.File
-import java.io.FileOutputStream
 import kotlin.coroutines.resume
 import kotlin.math.roundToInt
 import kotlin.random.Random
@@ -72,14 +59,30 @@ import kotlin.random.nextULong
 import kotlin.time.Duration
 
 internal object NtV2RichMediaSvc: BaseSvc() {
-    private const val GROUP_PIC_UPLOAD_TO = "100000000"
-
     private val requestIdSeq = atomic(2L)
+
+    private fun fetchGroupResUploadTo(): String {
+        return ShamrockConfig.getUpResGroup().ifEmpty { "100000000" }
+    }
+
+    suspend fun tryUploadResourceByNt(
+        chatType: Int,
+        elementType: Int,
+        resources: ArrayList<File>,
+        timeout: Duration,
+        retryCnt: Int = 5
+    ): Result<MutableList<CommonFileInfo>> {
+        return internalTryUploadResourceByNt(chatType, elementType, resources, timeout).onFailure {
+            if (retryCnt > 0) {
+                return tryUploadResourceByNt(chatType, elementType, resources, timeout, retryCnt - 1)
+            }
+        }
+    }
 
     /**
      * 批量上传图片
      */
-    suspend fun tryUploadResourceByNt(
+    private suspend fun internalTryUploadResourceByNt(
         chatType: Int,
         elementType: Int,
         resources: ArrayList<File>,
@@ -276,8 +279,9 @@ internal object NtV2RichMediaSvc: BaseSvc() {
         }
         val contact = when(chatType) {
             MsgConstant.KCHATTYPEC2C -> MessageHelper.generateContact(chatType, TicketSvc.getUin())
-            else -> Contact(chatType, GROUP_PIC_UPLOAD_TO, GROUP_PIC_UPLOAD_TO)
+            else -> Contact(chatType, fetchGroupResUploadTo(), null)
         }
+        LogCenter.log(contact.toString())
         val result = mutableListOf<CommonFileInfo>()
         withTimeoutOrNull(timeout) {
             suspendCancellableCoroutine {
@@ -297,10 +301,12 @@ internal object NtV2RichMediaSvc: BaseSvc() {
                     message = ArrayList(messages),
                     uniseq = uniseq.qqMsgId
                 ) { _, _ ->
-                    val kernelService = NTServiceFetcher.kernelService
-                    val sessionService = kernelService.wrapperSession
-                    val msgService = sessionService.msgService
-                    msgService.deleteMsg(contact, arrayListOf(uniseq.qqMsgId), null)
+                    if (contact.chatType == MsgConstant.KCHATTYPEGROUP && contact.peerUid == "100000000") {
+                        val kernelService = NTServiceFetcher.kernelService
+                        val sessionService = kernelService.wrapperSession
+                        val msgService = sessionService.msgService
+                        msgService.deleteMsg(contact, arrayListOf(uniseq.qqMsgId), null)
+                    }
                 }
                 it.invokeOnCancellation {
                     RichMediaUploadHandler.removeListener(uniseq.qqMsgId)
